Integration of GUI into game AB#685
